drop.tip {treeio}	R Documentation

drop.tip method

Description

drop.tip method

drop.tip method

Usage


drop.tip(object, tip, ...)

drop.tip(object, tip, ...)

## S4 method for signature 'phylo'
drop.tip(object, tip, ...)

Arguments

object
A treedata or phylo object

tip
a vector of mode numeric or character specifying the tips to delete

...
additional parameters

Value

updated object

Author(s)

Casey Dunn http://dunnlab.org and Guangchuang Yu https://guangchuangyu.github.io

Source

drop.tip for phylo object is a wrapper method of ape::drop.tip from the ape package. The documentation you should read for the drop.tip function can be found here: drop.tip

See Also

drop.tip

Examples


nhxfile <- system.file("extdata/NHX", "ADH.nhx", package="treeio")
nhx <- read.nhx(nhxfile)
drop.tip(nhx, c("ADH2", "ADH1"))

[Package treeio version 1.4.3 Index]