read.hyphy {treeio}	R Documentation

read.hyphy

Description

read HYPHY output

Usage


read.hyphy(nwk, ancseq, tip.fasfile = NULL)

Arguments

nwk
tree file in nwk format, one of hyphy output

ancseq
ancestral sequence file in nexus format, one of hyphy output

tip.fasfile
tip sequence file

Value

A hyphy object

Author(s)

Guangchuang Yu https://guangchuangyu.github.io

Examples


nwk <- system.file("extdata/HYPHY", "labelledtree.tree", package="treeio")
ancseq <- system.file("extdata/HYPHY", "ancseq.nex", package="treeio")
read.hyphy(nwk, ancseq)

[Package treeio version 1.6.2 Index]