read.mcmctree {treeio}	R Documentation

read.mcmctree

Description

read MCMCTree output Tree

Usage

read.mcmctree(file, force.ultrametric = FALSE)

Arguments

file
the output tree file of MCMCTree

force.ultrametric
logical whether convert the tree to be ultrametric, if it is not ultrametric, default is FALSE. When the tree is ultrametric, branch times will be calculated automatically.

Value

treedata object

Examples

file <- system.file("extdata/MCMCTree", "mcmctree_output.tree", package="treeio")
tr <- read.mcmctree(file)
tr

[Package treeio version 1.18.1 Index]