root.phylo {treeio}	R Documentation

root

Description

re-root a tree

Usage

## S3 method for class 'phylo'
root(phy, outgroup, node = NULL, resolve.root = TRUE, ...)

## S3 method for class 'treedata'
root(phy, outgroup, node = NULL, resolve.root = TRUE, ...)

Arguments

phy
tree object

outgroup
a vector of mode numeric or character specifying the new outgroup

node
node to reroot

resolve.root
a logical specifying whether to resolve the new root as a bifurcating node

...
additional parameters passed to ape::root.phylo

Value

rerooted tree

Author(s)

Guangchuang Yu

[Package treeio version 1.12.0 Index]